update Management code
the old version returns the minPayable function that does not work correctly, causing the test case to test less than minPayable should work in theory but the branch coverage is still 0 about Management. after updating the code of Management calculate the return value relevant code. the branch coverage be covered